Move Newtson code back to main json package spec reader
