+++ src/NuGet.Core/NuGet.ProjectModel/LockFile/LockFileFormat.cs
@@ -938,7 +938,7 @@ namespace NuGet.ProjectModel
                     NuGetFramework framework = NuGetFramework.Parse(frameworkPropertyName);
                     var dependencies = new List<LibraryDependency>();
 
-                    NjPackageSpecReader.ReadCentralTransitiveDependencyGroup(
+                    JsonPackageSpecReader.ReadCentralTransitiveDependencyGroup(
                         jsonReader: jsonReader,
                         results: dependencies,
                         packageSpecPath: path);
